# интерфейсы/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Кейс 02: регистрация, оплата и доступ",
    "description": "Исполняемая демонстрация проектируемого контракта. Не принимает настоящие платежи. Тренировочные данные API не принимает.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/payments": {
      "post": {
        "tags": [
          "Оплата"
        ],
        "summary": "Создать или повторить запрос оплаты",
        "operationId": "createPayment",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 64,
              "title": "Ключ повторяемости запроса"
            },
            "description": "Ключ повторяемости запроса"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentIn"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Запрос принят; результат следует уточнить",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentOut"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Некорректный запрос"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Необходим действующий сеанс"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Доступ запрещён"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Запись не найдена или недоступна"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Конфликт состояния или повторного запроса"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Недопустимые входные данные"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Превышена частота запросов"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Сервис временно недоступен"
          }
        }
      }
    },
    "/v1/payments/{payment_id}": {
      "get": {
        "tags": [
          "Оплата"
        ],
        "summary": "Получить состояние своего платежа",
        "operationId": "getPayment",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Идентификатор платежа"
            },
            "description": "Идентификатор платежа"
          }
        ],
        "responses": {
          "200": {
            "description": "Операция выполнена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentOut"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Некорректный запрос"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Необходим действующий сеанс"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Доступ запрещён"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Запись не найдена или недоступна"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Конфликт состояния или повторного запроса"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Недопустимые входные данные"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Превышена частота запросов"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Сервис временно недоступен"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Код",
            "description": "Код. Идентификатор поля в формате обмена: code."
          },
          "message": {
            "type": "string",
            "title": "Сообщение для пользователя",
            "description": "Сообщение для пользователя. Идентификатор поля в формате обмена: message."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Ошибка операции"
      },
...
      "PaymentIn": {
        "properties": {
          "planCode": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Код тарифа",
            "description": "Код тарифа. Идентификатор поля в формате обмена: planCode."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "planCode"
        ],
        "title": "Создание оплаты"
      },
      "PaymentOut": {
        "properties": {
          "paymentId": {
            "type": "string",
            "format": "uuid",
            "title": "Идентификатор платежа",
            "description": "Идентификатор платежа. Идентификатор поля в формате обмена: paymentId."
          },
          "status": {
            "type": "string",
            "enum": [
              "creating",
              "pending",
              "succeeded",
              "canceled",
              "refunded"
            ],
            "title": "Состояние",
            "description": "Состояние. Идентификатор поля в формате обмена: status. Допустимые обозначения: creating — Исход создания неизвестен; pending — Ожидает оплаты; succeeded — Оплата подтверждена; canceled — Платёж отменён; refunded — Полный возврат."
          },
          "amountMinor": {
            "type": "integer",
            "title": "Сумма в минимальных денежных единицах",
            "description": "Сумма в минимальных денежных единицах. Идентификатор поля в формате обмена: amountMinor."
          },
          "currency": {
            "type": "string",
            "title": "Код валюты",
            "description": "Код валюты. Идентификатор поля в формате обмена: currency."
          },
          "checkoutUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Адрес страницы оплаты",
            "description": "Адрес страницы оплаты. Идентификатор поля в формате обмена: checkoutUrl."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "paymentId",
          "status",
          "amountMinor",
          "currency",
          "checkoutUrl"
        ],
        "title": "Состояние платежа"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Ключ сеанса передаётся в заголовке Authorization: Bearer <ключ>. Значение не публикуется в журнале."
      }
    }
  }
}